+++ stdlib/LinearAlgebra/test/generic.jl
@@ -120,7 +120,7 @@ end
     @test [linreg(view(x,1:6),view(y,1:6))...] ≈ [3.8366666666666642,1.3271428571428574]
 
     # check (LinSpace, UnitRange)
-    x = linspace(1.0, 12.0, 100)
+    x = range(1.0, stop=12.0, length=100)
     y = -100:-1
     @test [linreg(x, y)...] ≈ [-109.0, 9.0]
 
@@ -130,8 +130,8 @@ end
     @test [linreg(x, y)...] ≈ [13.0, -1.0]
 
     # check (LinSpace, LinSpace)
-    x = linspace(-5, 10, 100)
-    y = linspace(50, 200, 100)
+    x = range(-5, stop=10, length=100)
+    y = range(50, stop=200, length=100)
     @test [linreg(x, y)...] ≈ [100.0, 10.0]
 
     # check (Array, Array)